Move skills modal code into skills report page, delete modal
