--- web-ui/src/components/skills/SkillsReportModal.css
@@ -1,37 +1,0 @@
-.SkillsReportModal {
-    position: absolute;
-    min-width: 400px;
-    max-width: 600px;
-    background-color: #FFF;
-    top: 50%;
-    left: 50%;
-    padding: .5rem;
-    transform: translate(-50%, -50%);
-    border: 2px solid #FFF;
-}
-
-.SkillsReportModal h2 {
-    margin-block-end: 0rem;
-    margin-left: .5rem;
-}
-
-.SkillsReportModal-actions {
-    margin-top: 1rem;
-    width: calc(100% - 1rem);
-    display: flex;
-    flex-direction: row;
-    justify-content: flex-end;
-}
-
-.SkillsReportModal .MuiTextField-root.fullWidth {
-    width: calc(100% - 1rem);
-}
-
-.SkillsReportModal .MuiTextField-root.halfWidth {
-    width: calc(50% - .5rem);
-}
-
-.SkillsReportModal .MuiTextField-root {
-    margin: .5rem;
-    width: 25ch;
-  }
